Module libafl::bolts::core_affinity

This crate manages CPU affinities.

Example

This example shows how create a thread for each available processor and pin each thread to its corresponding processor.

use libafl::bolts::core_affinity;

use std::thread;

// Retrieve the IDs of all active CPU cores.
let core_ids = core_affinity::get_core_ids().unwrap();

// Create a thread for each active CPU core.
let handles = core_ids.into_iter().map(|id| {
    thread::spawn(move || {
        // Pin this thread to a single CPU core.
        id.set_affinity();
        // Do more work after this.
    })
}).collect::<Vec<_>>();

for handle in handles.into_iter() {
    handle.join().unwrap();
}

This file is a fork of https://github.com/Elzair/core_affinity_rs

Structs

CoreId

This represents a CPU core.

Cores

A list of CoreId to use for fuzzing

Functions

get_core_ids

This function tries to retrieve information on all the “cores” active on this system.

parse_core_bind_argDeprecated

Parses core binding args from user input. Returns a Vec of CPU IDs.