Module libafl::bolts::os

Operating System specific abstractions

Modules

pipes

Unix pipe wrapper for LibAFL

unix_shmem_server

On Android, we can only share maps between processes by serializing fds over sockets. On MacOS, we cannot rely on reference counting for Maps. Hence, the unix_shmem_server keeps track of existing maps, creates new maps for clients, and forwards them over unix domain sockets.

unix_signals

Signal handling for unix

Structs

ChildHandle

Child Process Handle

Enums

ForkResult

The ForkResult (result of a fork)

Functions

dup2

“Safe” wrapper around dup2

fork⚠

Unix has forks.

startable_self

Executes the current process from the beginning, as subprocess. use start_self.status()? to wait for the child