Module libafl::bolts::os::unix_signals

Signal handling for unix

Structs

siginfo_t

ucontext_t

Enums

Signal

All signals on this system, as enum.

c_void

Equivalent to C’s void type when used as a pointer.

Statics

CRASH_SIGNALS

A list of crashing signals

Traits

Handler

A trait for LibAFL signal handling

Functions

setup_signal_handler⚠

Setup signal handlers in a somewhat rusty way. This will allocate a signal stack and set the signal handlers accordingly. It is, for example, used in the crate::executors::InProcessExecutor to restart the fuzzer in case of a crash, or to handle SIGINT in the broker process.

ucontext

Function to get the current ucontext_t for this process. This calls the libc getcontext function under the hood. It can be useful, for example for dump_regs. Note that calling this method may, of course, alter the state. We wrap it here, as it seems to be (currently) not available on MacOS in the libc crate.