Struct libafl::bolts::ownedref::OwnedSliceMut

pub struct OwnedSliceMut<'a, T: 'a + Sized> { /* private fields */ }

Wrap a mutable slice and convert to a Vec on serialize

Implementations

impl<'a, T: 'a + Sized> OwnedSliceMut<'a, T>

pub unsafe fn from_raw_parts_mut(ptr: *mut T, len: usize) -> OwnedSliceMut<'a, T>

Create a new OwnedSliceMut from a raw pointer and length

Safety

The pointer must be valid and point to a map of the size size_of<T>() * len The contents will be dereferenced in subsequent operations.