Module libafl::bolts::rands

The random number generators of LibAFL

Structs

Lehmer64Rand

XXH3 Based, hopefully speedy, rnd implementation

RomuDuoJrRand

see https://arxiv.org/pdf/2002.11331.pdf

RomuTrioRand

Extremely quick rand implementation see https://arxiv.org/pdf/2002.11331.pdf

XorShift64Rand

XXH3 Based, hopefully speedy, rnd implementation

Xoshiro256StarRand

XXH3 Based, hopefully speedy, rnd implementation

Traits

Rand

Ways to get random around here. Please note that these are not cryptographically secure. Or, even if some might be by accident, at least they are not seeded in a cryptographically secure fashion.

RandomSeed

Initialize Rand types from a source of randomness.

Type Definitions

StdRand

The standard rand implementation for LibAFL. It is usually the right choice, with very good speed and a reasonable randomness. Not cryptographically secure (which is not what you want during fuzzing ;) )