Type Definition libafl::bolts::rands::StdRand

pub type StdRand = RomuDuoJrRand;

The standard rand implementation for LibAFL. It is usually the right choice, with very good speed and a reasonable randomness. Not cryptographically secure (which is not what you want during fuzzing ;) )