Struct libafl::bolts::serdeany::serdeany_registry::NamedSerdeAnyMap

pub struct NamedSerdeAnyMap { /* private fields */ }

A serializable HashMap wrapper for SerdeAny types, addressable by name.

Implementations

impl NamedSerdeAnyMap

pub fn get<T>(&self, name: &str) -> Option<&T> where
T: SerdeAny,

Get an element by name

pub fn by_typeid(&self, name: &str, typeid: &TypeId) -> Option<&dyn SerdeAny>

Get an element of a given type contained in this map by TypeId.

pub fn get_mut<T>(&mut self, name: &str) -> Option<&mut T> where
T: SerdeAny,

Get an element of a given type contained in this map by TypeId, as mut.

pub fn by_typeid_mut(
&mut self,
name: &str,
typeid: &TypeId
) -> Option<&mut dyn SerdeAny>

Get an element of a given type contained in this map by TypeId, as mut.

pub fn get_all<T>(
&self
) -> Option<Map<Values<'_, u64, Box<dyn SerdeAny>>, fn(_: &Box<dyn SerdeAny>) -> &T>> where
T: SerdeAny,

Get all elements of a type contained in this map.

pub fn all_by_typeid(
&self,
typeid: &TypeId
) -> Option<Map<Values<'_, u64, Box<dyn SerdeAny>>, fn(_: &Box<dyn SerdeAny>) -> &dyn SerdeAny>>

Get all elements of a given type contained in this map by TypeId.

pub fn get_all_mut<T>(
&mut self
) -> Option<Map<ValuesMut<'_, u64, Box<dyn SerdeAny>>, fn(_: &mut Box<dyn SerdeAny>) -> &mut T>> where
T: SerdeAny,

Get all elements contained in this map, as mut.

pub fn all_by_typeid_mut(
&mut self,
typeid: &TypeId
) -> Option<Map<ValuesMut<'_, u64, Box<dyn SerdeAny>>, fn(_: &mut Box<dyn SerdeAny>) -> &mut dyn SerdeAny>>

Get all TypeIds contained in this map, as mut.

pub fn all_typeids(
&self
) -> Map<Keys<'_, u64, HashMap<u64, Box<dyn SerdeAny>>>, fn(_: &u64) -> TypeId>

Get all TypeIds contained in this map.

pub fn for_each<F: FnMut(&TypeId, &Box<dyn SerdeAny>) -> Result<(), Error>>(
&self,
func: &mut F
) -> Result<(), Error>

Run func for each element in this map.

pub fn for_each_mut<F: FnMut(&TypeId, &mut Box<dyn SerdeAny>) -> Result<(), Error>>(
&mut self,
func: &mut F
) -> Result<(), Error>

Run func for each element in this map, getting a mutable borrow.

pub fn insert<T>(&mut self, val: T, name: &str) where
T: SerdeAny,

Insert an element into this map.

pub fn len(&self) -> usize

Returns the len of this map.

pub fn is_empty(&self) -> bool

Returns true if this map is empty.

pub fn contains_type<T>(&self) -> bool where
T: SerdeAny,

Returns if the element with a given type is contained in this map.

pub fn contains<T>(&self, name: &str) -> bool where
T: SerdeAny,

Returns if the element by a given name is contained in this map.

pub fn new() -> Self

Create a new SerdeAny map.