Module libafl::bolts::shmem::unix_shmem

A Unix sharedmem implementation.

On Android, this is partially reused to wrap unix_shmem::ashmem::AshmemShMem, Although for an ServedShMemProvider using a unix domain socket Is needed on top.

Modules

ashmem

Module containing ashmem shared memory support, commonly used on Android.

Structs

MmapShMem

Mmap-based The sharedmap impl for unix using shm_open and mmap. Default on MacOS and iOS, where we need a central point to unmap shared mem segments for dubious Mach kernel reasons.

MmapShMemProvider

A ShMemProvider which uses shmget/shmat/shmctl to provide shared memory mappings.

Type Definitions

UnixShMem

Shared memory for Unix

UnixShMemProvider

Shared memory Provider for Unix