Module libafl::bolts::tuples

Compiletime lists/tuples used throughout the LibAFL universe

Macros

tuple_list

Macro creating tuple list values from list of expressions.

tuple_list_type

Macro creating tuple list types from list of element types.

Traits

Append

Append to a tuple

HasConstLen

Gets the length of the element

HasNameId

Finds the const_name and name_id

HasNameIdTuple

Gets the id and const_name for the given index in a tuple

MatchFirstType

Returns the first element with the given type

MatchName

Match for a name and return the value

MatchNameAndType

Finds an element of a type by the given name.

MatchType

Match by type

Merge

Merge two TupleList

Named

We need fixed names for many parts of this lib.

NamedTuple

A named tuple

Prepend

Allows prepending of values to a tuple

TupleList

Trait providing conversion from tuple list into tuple.

Functions

type_eq

Returns if the type T is equal to U From https://stackoverflow.com/a/60138532/7658998