Trait libafl::bolts::tuples::MatchType

pub trait MatchType {
    fn match_type<T: 'static, FN: FnMut(&T)>(&self, f: &mut FN);
    fn match_type_mut<T: 'static, FN: FnMut(&mut T)>(&mut self, f: &mut FN);
}

Match by type

Required Methods

fn match_type<T: 'static, FN: FnMut(&T)>(&self, f: &mut FN)

Match by type and call the passed f function with a borrow, if found

fn match_type_mut<T: 'static, FN: FnMut(&mut T)>(&mut self, f: &mut FN)

Match by type and call the passed f function with a mutable borrow, if found