Module libafl::corpus

Corpuses contain the testcases, either in memory, on disk, or somewhere else.

Re-exports

pub use testcase::SchedulerTestcaseMetaData;

pub use testcase::Testcase;

pub use inmemory::InMemoryCorpus;

pub use ondisk::OnDiskCorpus;

pub use cached::CachedOnDiskCorpus;

Modules

cached

The cached ondisk corpus stores testcases to disk keeping a part of them in memory.

inmemory

In-memory corpus, keeps all test cases in memory at all times

ondisk

The ondisk corpus stores unused testcases to disk.

testcase

The testcase is a struct embedded in each corpus. It will contain a respective input, and metadata.

Traits

Corpus

Corpus with all current testcases