Struct libafl::corpus::inmemory::InMemoryCorpus

pub struct InMemoryCorpus<I> where
    I: Input, { /* private fields */ }

A corpus handling all in memory.

Implementations

impl<I> InMemoryCorpus<I> where
I: Input,

pub fn new() -> Self

Creates a new InMemoryCorpus, keeping all Testcases in memory. This is the simplest and fastest option, however test progress will be lost on exit or on OOM.