Struct libafl::corpus::testcase::Testcase

pub struct Testcase<I> where
    I: Input, { /* private fields */ }

An entry in the Testcase Corpus

Implementations

impl<I> Testcase<I> where
I: Input,

Impl of a testcase

pub fn load_input(&mut self) -> Result<&I, Error>

Returns this testcase with a loaded input

pub fn store_input(&mut self) -> Result<bool, Error>

Store the input to disk if possible

pub fn input(&self) -> &Option<I>

Get the input, if any

pub fn input_mut(&mut self) -> &mut Option<I>

Get the input, if any (mutable)

pub fn set_input(&mut self, input: I)

Set the input

pub fn filename(&self) -> &Option<String>

Get the filename, if any

pub fn filename_mut(&mut self) -> &mut Option<String>

Get the filename, if any (mutable)

pub fn set_filename(&mut self, filename: String)

Set the filename

pub fn exec_time(&self) -> &Option<Duration>

Get the execution time of the testcase

pub fn exec_time_mut(&mut self) -> &mut Option<Duration>

Get the execution time of the testcase (mutable)

pub fn set_exec_time(&mut self, time: Duration)

Sets the execution time of the current testcase

pub fn executions(&self) -> &usize

Get the executions

pub fn executions_mut(&mut self) -> &mut usize

Get the executions (mutable)

pub fn fuzz_level(&self) -> usize

Get the fuzz_level

pub fn set_fuzz_leve(&mut self, fuzz_level: usize)

Set the fuzz_level

pub fn fuzzed(&self) -> bool

Get if it was fuzzed

pub fn set_fuzzed(&mut self, fuzzed: bool)

Set if it was fuzzed

pub fn new<T>(input: T) -> Self where
T: Into<I>,

Create a new Testcase instace given an input

pub fn with_filename(input: I, filename: String) -> Self

Create a new Testcase instance given an Input and a filename

pub fn with_executions(input: I, executions: usize) -> Self

Create a new Testcase instance given an Input and the number of executions

impl<I> Testcase<I> where
I: Input + HasLen,

Impl of a testcase when the input has len

pub fn cached_len(&mut self) -> Result<usize, Error>

Get the cached len