Trait libafl::corpus::Corpus

pub trait Corpus<I>: Serialize + for<'de> Deserialize<'de> where
    I: Input, {
    fn count(&self) -> usize;
    fn add(&mut self, testcase: Testcase<I>) -> Result<usize, Error>;
    fn replace(&mut self, idx: usize, testcase: Testcase<I>) -> Result<(), Error>;
    fn remove(&mut self, idx: usize) -> Result<Option<Testcase<I>>, Error>;
    fn get(&self, idx: usize) -> Result<&RefCell<Testcase<I>>, Error>;
    fn current(&self) -> &Option<usize>;
    fn current_mut(&mut self) -> &mut Option<usize>;

    fn is_empty(&self) -> bool { ... }
}

Corpus with all current testcases

Required Methods

fn count(&self) -> usize

Returns the number of elements

fn add(&mut self, testcase: Testcase<I>) -> Result<usize, Error>

Add an entry to the corpus and return its index

fn replace(&mut self, idx: usize, testcase: Testcase<I>) -> Result<(), Error>

Replaces the testcase at the given idx

fn remove(&mut self, idx: usize) -> Result<Option<Testcase<I>>, Error>

Removes an entry from the corpus, returning it if it was present.

fn get(&self, idx: usize) -> Result<&RefCell<Testcase<I>>, Error>

Get by id

fn current(&self) -> &Option<usize>

Current testcase scheduled

fn current_mut(&mut self) -> &mut Option<usize>

Current testcase scheduled (mutable)

Provided Methods

fn is_empty(&self) -> bool

Returns true, if no elements are in this corpus yet