Enum libafl::Error

pub enum Error {

    Serialize(String, ErrorBacktrace),
    Compression(ErrorBacktrace),
    File(Error, ErrorBacktrace),
    EmptyOptional(String, ErrorBacktrace),
    KeyNotFound(String, ErrorBacktrace),
    Empty(String, ErrorBacktrace),
    IteratorEnd(String, ErrorBacktrace),
    NotImplemented(String, ErrorBacktrace),
    IllegalState(String, ErrorBacktrace),
    IllegalArgument(String, ErrorBacktrace),
    Unsupported(String, ErrorBacktrace),
    ShuttingDown,
    Unknown(String, ErrorBacktrace),

}

Main error struct for LibAFL

Variants

Serialize(String, ErrorBacktrace)

Serialization error

Compression(ErrorBacktrace)

Compression error

File(Error, ErrorBacktrace)

File related error

EmptyOptional(String, ErrorBacktrace)

Optional val was supposed to be set, but isn’t.

KeyNotFound(String, ErrorBacktrace)

Key not in Map

Empty(String, ErrorBacktrace)

No elements in the current item

IteratorEnd(String, ErrorBacktrace)

End of iteration

NotImplemented(String, ErrorBacktrace)

This is not supported (yet)

IllegalState(String, ErrorBacktrace)

You’re holding it wrong

IllegalArgument(String, ErrorBacktrace)

The argument passed to this method or function is not valid

Unsupported(String, ErrorBacktrace)

The performed action is not supported on the current platform

ShuttingDown

Shutting down, not really an error.

Unknown(String, ErrorBacktrace)

Something else happened

Implementations

impl Error

pub fn serialize<S>(arg: S) -> Self where
S: Into<String>,

Serialization error

pub fn compression() -> Self

Compression error

pub fn file(arg: Error) -> Self

File related error

pub fn empty_optional<S>(arg: S) -> Self where
S: Into<String>,

Optional val was supposed to be set, but isn’t.

pub fn key_not_found<S>(arg: S) -> Self where
S: Into<String>,

Key not in Map

pub fn empty<S>(arg: S) -> Self where
S: Into<String>,

No elements in the current item

pub fn iterator_end<S>(arg: S) -> Self where
S: Into<String>,

End of iteration

pub fn not_implemented<S>(arg: S) -> Self where
S: Into<String>,

This is not supported (yet)

pub fn illegal_state<S>(arg: S) -> Self where
S: Into<String>,

You’re holding it wrong

pub fn illegal_argument<S>(arg: S) -> Self where
S: Into<String>,

The argument passed to this method or function is not valid

pub fn shutting_down() -> Self

Shutting down, not really an error.

pub fn unsupported<S>(arg: S) -> Self where
S: Into<String>,

This operation is not supported on the current architecture or platform

pub fn unknown<S>(arg: S) -> Self where
S: Into<String>,

Something else happened