Function libafl::events::llmp::setup_restarting_mgr_std

pub fn setup_restarting_mgr_std<I, MT, OT, S>(
    monitor: MT,
    broker_port: u16,
    configuration: EventConfig
) -> Result<(Option<S>, LlmpRestartingEventManager<I, OT, S, StdShMemProvider>), Error> where
    I: Input,
    MT: Monitor + Clone,
    OT: ObserversTuple<I, S> + DeserializeOwned,
    S: DeserializeOwned, 

Sets up a restarting fuzzer, using the StdShMemProvider, and standard features. The restarting mgr is a combination of restarter and runner, that can be used on systems with and without fork support. The restarter will spawn a new process each time the child crashes or timeouts.