Struct libafl::events::llmp::LlmpEventManager

pub struct LlmpEventManager<I, OT, S, SP> where
    I: Input,
    OT: ObserversTuple<I, S>,
    SP: ShMemProvider + 'static, { /* private fields */ }

An EventManager that forwards all events to other attached fuzzers on shared maps or via tcp, using low-level message passing, crate::bolts::llmp.

Implementations

impl<I, OT, S, SP> LlmpEventManager<I, OT, S, SP> where
I: Input,
OT: ObserversTuple<I, S>,
SP: ShMemProvider + 'static,

pub fn new(
llmp: LlmpClient<SP>,
configuration: EventConfig
) -> Result<Self, Error>

Create a manager from a raw llmp client

pub fn new_on_port(
shmem_provider: SP,
port: u16,
configuration: EventConfig
) -> Result<Self, Error>

Create llmp on a port If the port is not yet bound, it will act as broker Else, it will act as client.

pub fn existing_client_from_env(
shmem_provider: SP,
env_name: &str,
configuration: EventConfig
) -> Result<Self, Error>

If a client respawns, it may reuse the existing connection, previously stored by LlmpClient::to_env().

pub fn describe(&self) -> Result<LlmpClientDescription, Error>

Describe the client event mgr’s llmp parts in a restorable fashion

pub fn existing_client_from_description(
shmem_provider: SP,
description: &LlmpClientDescription,
configuration: EventConfig
) -> Result<Self, Error>

Create an existing client from description

pub fn to_env(&self, env_name: &str)

Write the config for a client EventManager to env vars, a new client can reattach using LlmpEventManager::existing_client_from_env().