Struct libafl::events::simple::SimpleRestartingEventManager

pub struct SimpleRestartingEventManager<I, MT, S, SP> where
    I: Input,
    SP: ShMemProvider,
    MT: Monitor + Debug, { /* private fields */ }

Provides a builder which can be used to build a SimpleRestartingEventManager, which is a combination of a restarter and runner, that can be used on systems both with and without fork support. The restarter will start a new process each time the child crashes or times out.

Implementations

impl<I, MT, S, SP> SimpleRestartingEventManager<I, MT, S, SP> where
I: Input,
SP: ShMemProvider,
MT: Monitor + Debug,

pub fn launch(
monitor: MT,
shmem_provider: &mut SP
) -> Result<(Option<S>, Self), Error> where
S: DeserializeOwned + Serialize + HasCorpus<I> + HasSolutions<I>,
MT: Debug,

Launch the simple restarting manager. This EventManager is simple and single threaded, but can still used shared maps to recover from crashes and timeouts.