Trait libafl::events::EventFirer

pub trait EventFirer<I> where
    I: Input, {
    fn fire<S>(&mut self, state: &mut S, event: Event<I>) -> Result<(), Error>;

    fn log<S>(
        &mut self,
        state: &mut S,
        severity_level: LogSeverity,
        message: String
    ) -> Result<(), Error> { ... }
    fn serialize_observers<OT, S>(
        &mut self,
        observers: &OT
    ) -> Result<Vec<u8>, Error>
    where
        OT: ObserversTuple<I, S> + Serialize,
    { ... }
    fn configuration(&self) -> EventConfig { ... }
}

EventFirer fire an event.

Required Methods

fn fire<S>(&mut self, state: &mut S, event: Event<I>) -> Result<(), Error>

Send off an Event to the broker

For multi-processed managers, such as llmp::LlmpEventManager, this serializes the Event and commits it to the llmp page. In this case, if you fire faster than the broker can consume (for example for each Input, on multiple cores) the llmp shared map may fill up and the client will eventually OOM or panic. This should not happen for a normal use-case.

Provided Methods

fn log<S>(
&mut self,
state: &mut S,
severity_level: LogSeverity,
message: String
) -> Result<(), Error>

Send off an Event::Log event to the broker. This is a shortcut for EventFirer::fire with Event::Log as argument.

fn serialize_observers<OT, S>(
&mut self,
observers: &OT
) -> Result<Vec<u8>, Error> where
OT: ObserversTuple<I, S> + Serialize,

Serialize all observers for this type and manager

fn configuration(&self) -> EventConfig

Get the configuration