Module libafl::executors::command

The command executor executes a sub program for each run

Structs

CommandExecutor

A CommandExecutor is a wrapper around std::process::Command to execute a target as a child process. Construct a CommandExecutor by implementing CommandConfigurator for a type of your choice and calling CommandConfigurator::into_executor on it. Instead, you can use CommandExecutor::builder() to construct a CommandExecutor backed by a StdCommandConfigurator.

CommandExecutorBuilder

The builder for a default CommandExecutor that should fit most use-cases.

StdCommandConfigurator

A simple Configurator that takes the most common parameters Writes the input either to stdio or to a file Use CommandExecutor::builder() to use this configurator.

Enums

InputLocation

How to deliver input to an external program StdIn: The traget reads from stdin File: The target reads from the specified InputFile

Traits

CommandConfigurator

A CommandConfigurator takes care of creating and spawning a std::process::Command for the CommandExecutor.