Module libafl::executors::forkserver

Expose an Executor based on a Forkserver in order to execute AFL/AFL++ binaries

Structs

Forkserver

The Forkserver is communication channel with a child process that forks on request of the fuzzer. The communication happens via pipe.

ForkserverExecutor

This Executor can run binaries compiled for AFL/AFL++ that make use of a forkserver. Shared memory feature is also available, but you have to set things up in your code. Please refer to AFL++’s docs. https://github.com/AFLplusplus/AFLplusplus/blob/stable/instrumentation/README.persistent_mode.md

ForkserverExecutorBuilder

The builder for ForkserverExecutor

TimeoutForkserverExecutor

The timeout forkserver executor that wraps around the standard forkserver executor and sets a timeout before each run.

Traits

ConfigTarget

Configure the target, limit, setsid, pipe_stdin, the code was borrowed from the Angora fuzzer

HasForkserver

A struct that has a forkserver