Struct libafl::executors::forkserver::ForkserverExecutorBuilder

pub struct ForkserverExecutorBuilder<'a, SP> { /* private fields */ }

The builder for ForkserverExecutor

Implementations

impl<'a, SP> ForkserverExecutorBuilder<'a, SP>

pub fn build<I, OT, S>(
&mut self,
observers: OT
) -> Result<ForkserverExecutor<I, OT, S, SP>, Error> where
I: Input + HasTargetBytes,
OT: ObserversTuple<I, S>,
SP: ShMemProvider,

Builds ForkserverExecutor.

pub fn autotokens(self, tokens: &'a mut Tokens) -> Self

Use autodict?

pub fn parse_afl_cmdline<IT, O>(self, args: IT) -> Self where
IT: IntoIterator<Item = O>,
O: AsRef<OsStr>,

Parse afl style command line

impl<'a> ForkserverExecutorBuilder<'a, StdShMemProvider>

pub fn new() -> ForkserverExecutorBuilder<'a, StdShMemProvider>

Creates a new AFL-style ForkserverExecutor with the given target, arguments and observers. This is the builder for ForkserverExecutor This Forkserver will attempt to provide inputs over shared mem when shmem_provider is given. Else this forkserver will try to write the input to .cur_input file. If debug_child is set, the child will print to stdout/stderr.

pub fn program<O>(self, program: O) -> Self where
O: AsRef<OsStr>,

The harness

pub fn arg<O>(self, arg: O) -> Self where
O: AsRef<OsStr>,

Adds an argument to the harness’s commandline

pub fn args<IT, O>(self, args: IT) -> Self where
IT: IntoIterator<Item = O>,
O: AsRef<OsStr>,

Adds arguments to the harness’s commandline

pub fn env<K, V>(self, key: K, val: V) -> Self where
K: AsRef<OsStr>,
V: AsRef<OsStr>,

Adds an environmental var to the harness’s commandline

pub fn envs<IT, K, V>(self, vars: IT) -> Self where
IT: IntoIterator<Item = (K, V)>,
K: AsRef<OsStr>,
V: AsRef<OsStr>,

Adds environmental vars to the harness’s commandline

pub fn arg_input_file<P: AsRef<Path>>(self, path: P) -> Self

Place the input at this position and set the filename for the input.

pub fn arg_input_file_std(self) -> Self

Place the input at this position and set the default filename for the input.

pub fn debug_child(self, debug_child: bool) -> Self

If debug_child is set, the child will print to stdout/stderr.

pub fn shmem_provider<SP: ShMemProvider>(
self,
shmem_provider: &'a mut SP
) -> ForkserverExecutorBuilder<'a, SP>

Shmem provider for forkserver’s shared memory testcase feature.