Trait libafl::executors::Executor

pub trait Executor<EM, I, S, Z>: Debug where
    I: Input, {
    fn run_target(
        &mut self,
        fuzzer: &mut Z,
        state: &mut S,
        mgr: &mut EM,
        input: &I
    ) -> Result<ExitKind, Error>;

    fn with_observers<OT>(self, observers: OT) -> WithObservers<Self, OT>
    where
        Self: Sized,
        OT: ObserversTuple<I, S>,
    { ... }
    fn post_run_reset(&mut self) { ... }
}

An executor takes the given inputs, and runs the harness/target.

Required Methods

fn run_target(
&mut self,
fuzzer: &mut Z,
state: &mut S,
mgr: &mut EM,
input: &I
) -> Result<ExitKind, Error>

Instruct the target about the input and run

Provided Methods

fn with_observers<OT>(self, observers: OT) -> WithObservers<Self, OT> where
Self: Sized,
OT: ObserversTuple<I, S>,

Wraps this Executor with the given ObserversTuple to implement HasObservers.

If the executor already implements HasObservers, then the original implementation will be overshadowed by the implementation of this wrapper.

fn post_run_reset(&mut self)

Custom Reset Handler, e.g., to reset timers