Module libafl::feedbacks::map

Map feedback, maximizing or minimizing maps, for example the afl-style map observer.

Structs

AllIsNovel

AllIsNovel consider everything a novelty. Here mostly just for debugging.

AndReducer

A AndReducer reduces the values returning the bitwise AND with the old value

DifferentIsNovel

Consider as novelty if the reduced value is different from the old value.

MapFeedback

The most common AFL-like feedback type

MapFeedbackMetadata

The state of MapFeedback

MapIndexesMetadata

A testcase metadata holding a list of indexes of a map

MapNoveltiesMetadata

A testcase metadata holding a list of indexes of a map

MaxReducer

A MaxReducer reduces int values and returns their maximum.

MinReducer

A MinReducer reduces int values and returns their minimum.

NextPow2IsNovel

Only consider as novel the values which are at least the next pow2 class of the old value

OneOrFilledIsNovel

A filter that only saves values which are at least the next pow2 class

OrReducer

A OrReducer reduces the values returning the bitwise OR with the old value

ReachabilityFeedback

A ReachabilityFeedback reports if a target has been reached.

Constants

MAPFEEDBACK_PREFIX

The prefix of the metadata names

Traits

IsNovel

A IsNovel function is used to discriminate if a reduced value is considered novel.

Reducer

A Reducer function is used to aggregate values for the novelty search

Type Definitions

AflMapFeedback

A MapFeedback that implements the AFL algorithm using an OrReducer combining the bits for the history map and the bit from HitcountsMapObserver.

MaxMapFeedback

A MapFeedback that strives to maximize the map contents.

MaxMapOneOrFilledFeedback

A MapFeedback that strives to maximize the map contents, but only, if a value is larger than pow2 of the previous.

MaxMapPow2Feedback

A MapFeedback that strives to maximize the map contents, but only, if a value is larger than pow2 of the previous.

MinMapFeedback

A MapFeedback that strives to minimize the map contents.