Trait libafl::feedbacks::map::IsNovel

pub trait IsNovel<T>: 'static + Debug where
    T: Default + Copy + 'static, {
    fn is_novel(old: T, new: T) -> bool;
}

A IsNovel function is used to discriminate if a reduced value is considered novel.

Required Methods

fn is_novel(old: T, new: T) -> bool

If a new value in the MapFeedback was found, this filter can decide if the result is considered novel or not.