Trait libafl::feedbacks::map::Reducer

pub trait Reducer<T>: 'static + Debug where
    T: Default + Copy + 'static, {
    fn reduce(first: T, second: T) -> T;
}

A Reducer function is used to aggregate values for the novelty search

Required Methods

fn reduce(first: T, second: T) -> T

Reduce two values to one value, with the current Reducer.