Trait libafl::feedbacks::Feedback

pub trait Feedback<I, S>: Named + Debug where
    I: Input,
    S: HasClientPerfMonitor, {
    fn is_interesting<EM, OT>(
        &mut self,
        state: &mut S,
        manager: &mut EM,
        input: &I,
        observers: &OT,
        exit_kind: &ExitKind
    ) -> Result<bool, Error>
    where
        EM: EventFirer<I>,
        OT: ObserversTuple<I, S>;

    fn init_state(&mut self, _state: &mut S) -> Result<(), Error> { ... }
    fn is_interesting_introspection<EM, OT>(
        &mut self,
        state: &mut S,
        manager: &mut EM,
        input: &I,
        observers: &OT,
        exit_kind: &ExitKind
    ) -> Result<bool, Error>
    where
        EM: EventFirer<I>,
        OT: ObserversTuple<I, S>,
    { ... }
    fn append_metadata(
        &mut self,
        _state: &mut S,
        _testcase: &mut Testcase<I>
    ) -> Result<(), Error> { ... }
    fn discard_metadata(
        &mut self,
        _state: &mut S,
        _input: &I
    ) -> Result<(), Error> { ... }
}

Feedbacks evaluate the observers. Basically, they reduce the information provided by an observer to a value, indicating the “interestingness” of the last run.

Required Methods

fn is_interesting<EM, OT>(
&mut self,
state: &mut S,
manager: &mut EM,
input: &I,
observers: &OT,
exit_kind: &ExitKind
) -> Result<bool, Error> where
EM: EventFirer<I>,
OT: ObserversTuple<I, S>,

is_interesting return if an input is worth the addition to the corpus

Provided Methods

fn init_state(&mut self, _state: &mut S) -> Result<(), Error>

Initializes the feedback state. This method is called after that the State is created.

fn is_interesting_introspection<EM, OT>(
&mut self,
state: &mut S,
manager: &mut EM,
input: &I,
observers: &OT,
exit_kind: &ExitKind
) -> Result<bool, Error> where
EM: EventFirer<I>,
OT: ObserversTuple<I, S>,

Returns if the result of a run is interesting and the value input should be stored in a corpus. It also keeps track of introspection stats.

fn append_metadata(
&mut self,
_state: &mut S,
_testcase: &mut Testcase<I>
) -> Result<(), Error>

Append to the testcase the generated metadata in case of a new corpus item

fn discard_metadata(&mut self, _state: &mut S, _input: &I) -> Result<(), Error>

Discard the stored metadata in case that the testcase is not added to the corpus