Trait libafl::fuzzer::EvaluatorObservers

pub trait EvaluatorObservers<I, OT, S>: Sized where
    I: Input,
    OT: ObserversTuple<I, S>, {
    fn evaluate_input_with_observers<E, EM>(
        &mut self,
        state: &mut S,
        executor: &mut E,
        manager: &mut EM,
        input: I,
        send_events: bool
    ) -> Result<(ExecuteInputResult, Option<usize>), Error>
    where
        E: Executor<EM, I, S, Self> + HasObservers<I, OT, S>,
        EM: EventManager<E, I, S, Self>;
}

Evaluate an input modifying the state of the fuzzer

Required Methods

fn evaluate_input_with_observers<E, EM>(
&mut self,
state: &mut S,
executor: &mut E,
manager: &mut EM,
input: I,
send_events: bool
) -> Result<(ExecuteInputResult, Option<usize>), Error> where
E: Executor<EM, I, S, Self> + HasObservers<I, OT, S>,
EM: EventManager<E, I, S, Self>,

Runs the input and triggers observers and feedback, returns if is interesting an (option) the index of the new testcase in the corpus