Trait libafl::fuzzer::ExecutionProcessor

pub trait ExecutionProcessor<I, OT, S> where
    OT: ObserversTuple<I, S>,
    I: Input, {
    fn process_execution<EM>(
        &mut self,
        state: &mut S,
        manager: &mut EM,
        input: I,
        observers: &OT,
        exit_kind: &ExitKind,
        send_events: bool
    ) -> Result<(ExecuteInputResult, Option<usize>), Error>
    where
        EM: EventFirer<I>;
}

Evaluate if an input is interesting using the feedback

Required Methods

fn process_execution<EM>(
&mut self,
state: &mut S,
manager: &mut EM,
input: I,
observers: &OT,
exit_kind: &ExitKind,
send_events: bool
) -> Result<(ExecuteInputResult, Option<usize>), Error> where
EM: EventFirer<I>,

Evaluate if a set of observation channels has an interesting state