Trait libafl::inputs::Input

pub trait Input: Clone + Serialize + DeserializeOwned + Debug {
    fn generate_name(&self, idx: usize) -> String;

    fn to_file<P>(&self, path: P) -> Result<(), Error>
    where
        P: AsRef<Path>,
    { ... }
    fn from_file<P>(path: P) -> Result<Self, Error>
    where
        P: AsRef<Path>,
    { ... }
    fn wrapped_as_testcase(&mut self) { ... }
}

An input for the target

Required Methods

fn generate_name(&self, idx: usize) -> String

Generate a name for this input

Provided Methods

fn to_file<P>(&self, path: P) -> Result<(), Error> where
P: AsRef<Path>,

Write this input to the file

fn from_file<P>(path: P) -> Result<Self, Error> where
P: AsRef<Path>,

Load the content of this input from a file

fn wrapped_as_testcase(&mut self)

An hook executed if the input is stored as Testcase