Module libafl::monitors

Keep stats, and display them to the user. Usually used in a broker, or main node, of some sort.

Re-exports

pub use multi::MultiMonitor;

pub use disk::OnDiskTOMLMonitor;

Modules

disk

Monitors that wrap a base one and log on disk

multi

Monitor to display both cumulative and per-client monitor

tui

Monitor based on tui-rs

Structs

ClientPerfMonitor

Client performance statistics

ClientStats

A simple struct to keep track of client monitor

NopMonitor

Monitor that print exactly nothing. Not good for debuging, very good for speed.

SimpleMonitor

Tracking monitor during fuzzing.

Enums

PerfFeature

Various features that are measured for performance

UserStats

User-defined stat types TODO define aggregation function (avg, median, max, …)

Constants

NUM_PERF_FEATURES

Number of features we can measure for performance

Traits

Monitor

The monitor trait keeps track of all the client’s monitor, and offers methods to display them.