Struct libafl::monitors::ClientStats

pub struct ClientStats {
    pub corpus_size: u64,
    pub executions: u64,
    pub objective_size: u64,
    pub last_window_time: Duration,
    pub user_monitor: HashMap<String, UserStats>,
    pub introspection_monitor: ClientPerfMonitor,
}

A simple struct to keep track of client monitor

Fields
corpus_size: u64

The corpus size for this client

executions: u64

The total executions for this client

objective_size: u64

The size of the objectives corpus for this client

last_window_time: Duration

The last time we got this information

user_monitor: HashMap<String, UserStats>

User-defined monitor

introspection_monitor: ClientPerfMonitor

Client performance statistics

Implementations

impl ClientStats

pub fn update_executions(&mut self, executions: u64, _cur_time: Duration)

We got a new information about executions for this client, insert them.

pub fn update_corpus_size(&mut self, corpus_size: u64)

We got a new information about corpus size for this client, insert them.

pub fn update_objective_size(&mut self, objective_size: u64)

We got a new information about objective corpus size for this client, insert them.

pub fn execs_per_sec(&mut self, cur_time: Duration) -> u64

Get the calculated executions per second for this client

pub fn update_user_stats(&mut self, name: String, value: UserStats)

Update the user-defined stat with name and value

pub fn get_user_stats(&mut self, name: &str) -> Option<&UserStats>

Get a user-defined stat using the name

pub fn update_introspection_monitor(
&mut self,
introspection_monitor: ClientPerfMonitor
)

Update the current ClientPerfMonitor with the given ClientPerfMonitor