Trait libafl::monitors::Monitor

pub trait Monitor {
    fn client_stats_mut(&mut self) -> &mut Vec<ClientStats>;
    fn client_stats(&self) -> &[ClientStats];
    fn start_time(&mut self) -> Duration;
    fn display(&mut self, event_msg: String, sender_id: u32);

    fn corpus_size(&self) -> u64 { ... }
    fn objective_size(&self) -> u64 { ... }
    fn total_execs(&mut self) -> u64 { ... }
    fn execs_per_sec(&mut self) -> u64 { ... }
    fn client_stats_mut_for(&mut self, client_id: u32) -> &mut ClientStats { ... }
}

The monitor trait keeps track of all the client’s monitor, and offers methods to display them.

Required Methods

fn client_stats_mut(&mut self) -> &mut Vec<ClientStats>

The client monitor (mutable)

fn client_stats(&self) -> &[ClientStats]

The client monitor

fn start_time(&mut self) -> Duration

Creation time

fn display(&mut self, event_msg: String, sender_id: u32)

Show the monitor to the user

Provided Methods

fn corpus_size(&self) -> u64

Amount of elements in the corpus (combined for all children)

fn objective_size(&self) -> u64

Amount of elements in the objectives (combined for all children)

fn total_execs(&mut self) -> u64

Total executions

fn execs_per_sec(&mut self) -> u64

Executions per second

fn client_stats_mut_for(&mut self, client_id: u32) -> &mut ClientStats

The client monitor for a specific id, creating new if it doesn’t exist