Module libafl::mutators

Mutators mutate input during fuzzing.

Re-exports

pub use scheduled::*;

pub use mutations::*;

pub use token_mutations::*;

pub use encoded_mutations::*;

pub use mopt_mutator::*;

pub use gramatron::*;

pub use grimoire::*;

Modules

encoded_mutations

Mutations for EncodedInputs

gramatron

Gramatron is the rewritten gramatron fuzzer in rust. See the original gramatron repo Gramatron for more details.

grimoire

Grimoire is the rewritten grimoire mutator in rust. See the original repo Grimoire for more details.

mopt_mutator

The MOpt mutator scheduler, see https://github.com/puppet-meteor/MOpt-AFL and https://www.usenix.org/conference/usenixsecurity19/presentation/lyu

mutations

A wide variety of mutations used during fuzzing.

scheduled

The ScheduledMutator schedules multiple mutations internally.

token_mutations

Tokens are what AFL calls extras or dictionaries. They may be inserted as part of mutations during fuzzing.

Enums

MutationResult

The result of a mutation. If the mutation got skipped, the target will not be executed with the returned input.

Traits

Mutator

A mutator takes input, and mutates it. Simple as that.

MutatorsTuple

A Tuple of Mutators that can execute multiple Mutators in a row.