Module libafl::mutators::encoded_mutations

Mutations for EncodedInputs

Structs

EncodedAddMutator

Adds or subtracts a random value up to ARITH_MAX to a random place in the codes Vec.

EncodedCopyMutator

Codes copy mutation for encoded inputs

EncodedCrossoverInsertMutator

Crossover insert mutation for encoded inputs

EncodedCrossoverReplaceMutator

Crossover replace mutation for encoded inputs

EncodedDecMutator

Decrement a random code in the input

EncodedDeleteMutator

Codes delete mutation for encoded inputs

EncodedIncMutator

Increment a random code in the input

EncodedInsertCopyMutator

Insert mutation for encoded inputs

EncodedRandMutator

Set a code in the input as a random value

Functions

encoded_mutations

Get the mutations that compose the encoded mutator