Module libafl::mutators::mutations

A wide variety of mutations used during fuzzing.

Structs

BitFlipMutator

Bitflip mutation for inputs with a bytes vector

ByteAddMutator

Adds or subtracts a random value up to ARITH_MAX to a [<$size>] at a random place in the Vec, in random byte order.

ByteDecMutator

Byte decrement mutation for inputs with a bytes vector

ByteFlipMutator

Byteflip mutation for inputs with a bytes vector

ByteIncMutator

Byte increment mutation for inputs with a bytes vector

ByteInterestingMutator

Inserts an interesting value at a random place in the input vector

ByteNegMutator

Byte negate mutation for inputs with a bytes vector

ByteRandMutator

Byte random mutation for inputs with a bytes vector

BytesCopyMutator

Bytes copy mutation for inputs with a bytes vector

BytesDeleteMutator

Bytes delete mutation for inputs with a bytes vector

BytesExpandMutator

Bytes expand mutation for inputs with a bytes vector

BytesInsertCopyMutator

Bytes insert and self copy mutation for inputs with a bytes vector

BytesInsertMutator

Bytes insert mutation for inputs with a bytes vector

BytesRandInsertMutator

Bytes random insert mutation for inputs with a bytes vector

BytesRandSetMutator

Bytes random set mutation for inputs with a bytes vector

BytesSetMutator

Bytes set mutation for inputs with a bytes vector

BytesSwapMutator

Bytes swap mutation for inputs with a bytes vector

CrossoverInsertMutator

Crossover insert mutation for inputs with a bytes vector

CrossoverReplaceMutator

Crossover replace mutation for inputs with a bytes vector

DwordAddMutator

Adds or subtracts a random value up to ARITH_MAX to a [<$size>] at a random place in the Vec, in random byte order.

DwordInterestingMutator

Inserts an interesting value at a random place in the input vector

QwordAddMutator

Adds or subtracts a random value up to ARITH_MAX to a [<$size>] at a random place in the Vec, in random byte order.

SpliceMutator

Splice mutation for inputs with a bytes vector

WordAddMutator

Adds or subtracts a random value up to ARITH_MAX to a [<$size>] at a random place in the Vec, in random byte order.

WordInterestingMutator

Inserts an interesting value at a random place in the input vector

Constants

ARITH_MAX

The max value that will be added or subtracted during add mutations

INTERESTING_8

Interesting 8-bit values from AFL

INTERESTING_16

Interesting 16-bit values from AFL

INTERESTING_32

Interesting 32-bit values from AFL

Functions

buffer_copy

Mem move between vecs

buffer_self_copy

Mem move in the own vec

buffer_set

A simple way to set buffer contents. The compiler does the heavy lifting. see https://stackoverflow.com/a/51732799/1345238/

str_decode

Decodes a dictionary token: ‘foo\x41\and"bar’ -> ‘fooA\and“bar’