Module libafl::mutators::scheduled

The ScheduledMutator schedules multiple mutations internally.

Re-exports

pub use crate::mutators::mutations::*;

pub use crate::mutators::token_mutations::*;

Structs

LogMutationMetadata

The metadata placed in a crate::corpus::Testcase by a LoggerScheduledMutator.

LoggerScheduledMutator

A logging Mutator that wraps around a StdScheduledMutator.

StdScheduledMutator

A Mutator that schedules one of the embedded mutations on each call.

Traits

ComposedByMutations

A Mutator that composes multiple mutations into one.

ScheduledMutator

A Mutator scheduling multiple Mutators for an input.

Functions

havoc_mutations

Get the mutations that compose the Havoc mutator

tokens_mutations

Get the mutations that uses the Tokens metadata

Type Definitions

HavocMutationsType

Tuple type of the mutations that compose the Havoc mutator