Module libafl::observers

Observers give insights about runs of a target, such as coverage, timing, stack depth, and more.

Re-exports

pub use map::*;

pub use cmp::*;

pub use stdio::StdErrObserver;

pub use stdio::StdOutObserver;

pub use stacktrace::*;

Modules

cmp

The CmpObserver provides access to the logged values of CMP instructions

concolic

Concolic Tracing

map

The MapObserver provides access a map, usually injected into the target

stacktrace

the StacktraceObserver looks up the stacktrace on the execution thread and computes a hash for it for dedupe

stdio

The StdOutObserver and StdErrObserver observers look at the stdout of a program The executor must explicitely support these observers. For example, they are supported on the crate::executors::CommandExecutor.

Structs

ListObserver

A simple observer with a list of things.

TimeObserver

A simple observer, just overlooking the runtime of the target.

Traits

Observer

Observers observe different information about the target. They can then be used by various sorts of feedback.

ObserverWithHashField

A trait for obervers with a hash field

ObserversTuple

A haskell-style tuple of observers