Trait libafl::observers::cmp::CmpObserver

pub trait CmpObserver<CM, I, S>: Observer<I, S> where
    CM: CmpMap, {
    fn usable_count(&self) -> usize;
    fn cmp_map(&self) -> &CM;
    fn cmp_map_mut(&mut self) -> &mut CM;

    fn add_cmpvalues_meta(&mut self, state: &mut S)
    where
        S: HasMetadata,
    { ... }
}

A CmpObserver observes the traced comparisons during the current execution using a CmpMap

Required Methods

fn usable_count(&self) -> usize

Get the number of usable cmps (all by default)

fn cmp_map(&self) -> &CM

Get the CmpMap

fn cmp_map_mut(&mut self) -> &mut CM

Get the CmpMap (mutable)

Provided Methods

fn add_cmpvalues_meta(&mut self, state: &mut S) where
S: HasMetadata,

Add CmpValuesMetadata to the State including the logged values. This routine does a basic loop filtering because loop index cmps are not interesting.