Enum libafl::observers::concolic::SymExpr

pub enum SymExpr {

    InputByte {
        offset: usize,
    },
    Integer {
        value: u64,
        bits: u8,
    },
    Integer128 {
        high: u64,
        low: u64,
    },
    Float {
        value: f64,
        is_double: bool,
    },
    NullPointer,
    True,
    False,
    Bool {
        value: bool,
    },
    Neg {
        op: SymExprRef,
    },
    Add {
        a: SymExprRef,
        b: SymExprRef,
    },
    Sub {
        a: SymExprRef,
        b: SymExprRef,
    },
    Mul {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedDiv {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedDiv {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedRem {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedRem {
        a: SymExprRef,
        b: SymExprRef,
    },
    ShiftLeft {
        a: SymExprRef,
        b: SymExprRef,
    },
    LogicalShiftRight {
        a: SymExprRef,
        b: SymExprRef,
    },
    ArithmeticShiftRight {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedLessThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedLessEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedGreaterThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    SignedGreaterEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedLessThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedLessEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedGreaterThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    UnsignedGreaterEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    Not {
        op: SymExprRef,
    },
    Equal {
        a: SymExprRef,
        b: SymExprRef,
    },
    NotEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    BoolAnd {
        a: SymExprRef,
        b: SymExprRef,
    },
    BoolOr {
        a: SymExprRef,
        b: SymExprRef,
    },
    BoolXor {
        a: SymExprRef,
        b: SymExprRef,
    },
    And {
        a: SymExprRef,
        b: SymExprRef,
    },
    Or {
        a: SymExprRef,
        b: SymExprRef,
    },
    Xor {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrdered {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedGreaterThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedGreaterEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedLessThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedLessEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatOrderedNotEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnordered {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedGreaterThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedGreaterEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedLessThan {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedLessEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatUnorderedNotEqual {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatAbs {
        op: SymExprRef,
    },
    FloatAdd {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatSub {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatMul {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatDiv {
        a: SymExprRef,
        b: SymExprRef,
    },
    FloatRem {
        a: SymExprRef,
        b: SymExprRef,
    },
    Sext {
        op: SymExprRef,
        bits: u8,
    },
    Zext {
        op: SymExprRef,
        bits: u8,
    },
    Trunc {
        op: SymExprRef,
        bits: u8,
    },
    IntToFloat {
        op: SymExprRef,
        is_double: bool,
        is_signed: bool,
    },
    FloatToFloat {
        op: SymExprRef,
        to_double: bool,
    },
    BitsToFloat {
        op: SymExprRef,
        to_double: bool,
    },
    FloatToBits {
        op: SymExprRef,
    },
    FloatToSignedInteger {
        op: SymExprRef,
        bits: u8,
    },
    FloatToUnsignedInteger {
        op: SymExprRef,
        bits: u8,
    },
    BoolToBits {
        op: SymExprRef,
        bits: u8,
    },
    Concat {
        a: SymExprRef,
        b: SymExprRef,
    },
    Extract {
        op: SymExprRef,
        first_bit: usize,
        last_bit: usize,
    },
    Insert {
        target: SymExprRef,
        to_insert: SymExprRef,
        offset: u64,
        little_endian: bool,
    },
    PathConstraint {
        constraint: SymExprRef,
        taken: bool,
        location: Location,
    },
    ExpressionsUnreachable {
        exprs: Vec<SymExprRef>,
    },
    Call {
        location: Location,
    },
    Return {
        location: Location,
    },
    BasicBlock {
        location: Location,
    },

}

SymExpr represents a message in the serialization format. The messages in the format are a perfect mirror of the methods that are called on the runtime during execution.

Variants

InputByte

Fields

offset: usize

Integer

Fields

value: u64

bits: u8

Integer128

Fields

high: u64

low: u64

Float

Fields

value: f64

is_double: bool

NullPointer

True

False

Bool

Fields

value: bool

Neg

Fields

op: SymExprRef

Add

Fields

a: SymExprRef

b: SymExprRef

Sub

Fields

a: SymExprRef

b: SymExprRef

Mul

Fields

a: SymExprRef

b: SymExprRef

UnsignedDiv

Fields

a: SymExprRef

b: SymExprRef

SignedDiv

Fields

a: SymExprRef

b: SymExprRef

UnsignedRem

Fields

a: SymExprRef

b: SymExprRef

SignedRem

Fields

a: SymExprRef

b: SymExprRef

ShiftLeft

Fields

a: SymExprRef

b: SymExprRef

LogicalShiftRight

Fields

a: SymExprRef

b: SymExprRef

ArithmeticShiftRight

Fields

a: SymExprRef

b: SymExprRef

SignedLessThan

Fields

a: SymExprRef

b: SymExprRef

SignedLessEqual

Fields

a: SymExprRef

b: SymExprRef

SignedGreaterThan

Fields

a: SymExprRef

b: SymExprRef

SignedGreaterEqual

Fields

a: SymExprRef

b: SymExprRef

UnsignedLessThan

Fields

a: SymExprRef

b: SymExprRef

UnsignedLessEqual

Fields

a: SymExprRef

b: SymExprRef

UnsignedGreaterThan

Fields

a: SymExprRef

b: SymExprRef

UnsignedGreaterEqual

Fields

a: SymExprRef

b: SymExprRef

Not

Fields

op: SymExprRef

Equal

Fields

a: SymExprRef

b: SymExprRef

NotEqual

Fields

a: SymExprRef

b: SymExprRef

BoolAnd

Fields

a: SymExprRef

b: SymExprRef

BoolOr

Fields

a: SymExprRef

b: SymExprRef

BoolXor

Fields

a: SymExprRef

b: SymExprRef

And

Fields

a: SymExprRef

b: SymExprRef

Or

Fields

a: SymExprRef

b: SymExprRef

Xor

Fields

a: SymExprRef

b: SymExprRef

FloatOrdered

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedGreaterThan

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedGreaterEqual

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedLessThan

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedLessEqual

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedEqual

Fields

a: SymExprRef

b: SymExprRef

FloatOrderedNotEqual

Fields

a: SymExprRef

b: SymExprRef

FloatUnordered

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedGreaterThan

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedGreaterEqual

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedLessThan

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedLessEqual

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedEqual

Fields

a: SymExprRef

b: SymExprRef

FloatUnorderedNotEqual

Fields

a: SymExprRef

b: SymExprRef

FloatAbs

Fields

op: SymExprRef

FloatAdd

Fields

a: SymExprRef

b: SymExprRef

FloatSub

Fields

a: SymExprRef

b: SymExprRef

FloatMul

Fields

a: SymExprRef

b: SymExprRef

FloatDiv

Fields

a: SymExprRef

b: SymExprRef

FloatRem

Fields

a: SymExprRef

b: SymExprRef

Sext

Fields

op: SymExprRef

bits: u8

Zext

Fields

op: SymExprRef

bits: u8

Trunc

Fields

op: SymExprRef

bits: u8

IntToFloat

Fields

op: SymExprRef

is_double: bool

is_signed: bool

FloatToFloat

Fields

op: SymExprRef

to_double: bool

BitsToFloat

Fields

op: SymExprRef

to_double: bool

FloatToBits

Fields

op: SymExprRef

FloatToSignedInteger

Fields

op: SymExprRef

bits: u8

FloatToUnsignedInteger

Fields

op: SymExprRef

bits: u8

BoolToBits

Fields

op: SymExprRef

bits: u8

Concat

Fields

a: SymExprRef

b: SymExprRef

Extract

Fields

op: SymExprRef

first_bit: usize

last_bit: usize

Insert

Fields

target: SymExprRef

to_insert: SymExprRef

offset: u64

little_endian: bool

PathConstraint

Fields

constraint: SymExprRef

taken: bool

location: Location

ExpressionsUnreachable

Fields

exprs: Vec<SymExprRef>

These expressions won’t be referenced again

Call

Fields

location: Location

Location information regarding a call. Tracing this information is optional.

Return

Fields

location: Location

Location information regarding a return. Tracing this information is optional.

BasicBlock

Fields

location: Location

Location information regarding a basic block. Tracing this information is optional.