Module libafl::observers::map

The MapObserver provides access a map, usually injected into the target

Structs

ConstMapObserver

Use a const size to speedup Feedback::is_interesting when the user can know the size of the map at compile time.

HitcountsMapObserver

Map observer with hitcounts postprocessing

MapObserverSimpleIteratoMut

A Simple iterator calling MapObserver::get_mut

MapObserverSimpleIterator

A Simple iterator calling MapObserver::get

MultiMapObserver

The Multi Map Observer merge different maps into one observer

OwnedMapObserver

Exact copy of StdMapObserver that owns its map Used for python bindings

StdMapObserver

The Map Observer retrieves the state of a map, that will get updated by the target. A well-known example is the AFL-Style coverage map.

VariableMapObserver

Overlooking a variable bitmap

Traits

MapObserver

A MapObserver observes the static map, as oftentimes used for AFL-like coverage information