Trait libafl::observers::map::MapObserver

pub trait MapObserver: HasLen + Named + Serialize + DeserializeOwned + Debug {
    type Entry: Bounded + PartialEq + Default + Copy + Debug + 'static;

    fn get(&self, idx: usize) -> &Self::Entry;
    fn get_mut(&mut self, idx: usize) -> &mut Self::Entry;
    fn usable_count(&self) -> usize;
    fn count_bytes(&self) -> u64;
    fn hash(&self) -> u64;
    fn initial(&self) -> Self::Entry;
    fn initial_mut(&mut self) -> &mut Self::Entry;
    fn reset_map(&mut self) -> Result<(), Error>;
    fn to_vec(&self) -> Vec<Self::Entry>;
    fn how_many_set(&self, indexes: &[usize]) -> usize;

    fn set_initial(&mut self, initial: Self::Entry) { ... }
}

A MapObserver observes the static map, as oftentimes used for AFL-like coverage information

TODO: enforce iter() -> AssociatedTypeIter when generic associated types stabilize

Required Associated Types

type Entry: Bounded + PartialEq + Default + Copy + Debug + 'static

Type of each entry in this map

Required Methods

fn get(&self, idx: usize) -> &Self::Entry

Get the value at idx

fn get_mut(&mut self, idx: usize) -> &mut Self::Entry

Get the value at idx (mutable)

fn usable_count(&self) -> usize

Get the number of usable entries in the map (all by default)

fn count_bytes(&self) -> u64

Count the set bytes in the map

fn hash(&self) -> u64

Compute the hash of the map

fn initial(&self) -> Self::Entry

Get the initial value for reset()

fn initial_mut(&mut self) -> &mut Self::Entry

Get the initial value for reset() (mutable)

fn reset_map(&mut self) -> Result<(), Error>

Reset the map

fn to_vec(&self) -> Vec<Self::Entry>

Get these observer’s contents as Vec

fn how_many_set(&self, indexes: &[usize]) -> usize

Get the number of set entries with the specified indexes

Provided Methods

fn set_initial(&mut self, initial: Self::Entry)

Set the initial value for reset()