Trait libafl::observers::Observer

pub trait Observer<I, S>: Named + Debug {
    fn flush(&mut self) -> Result<(), Error> { ... }
    fn pre_exec(&mut self, _state: &mut S, _input: &I) -> Result<(), Error> { ... }
    fn post_exec(
        &mut self,
        _state: &mut S,
        _input: &I,
        _exit_kind: &ExitKind
    ) -> Result<(), Error> { ... }
    fn pre_exec_child(&mut self, _state: &mut S, _input: &I) -> Result<(), Error> { ... }
    fn post_exec_child(
        &mut self,
        _state: &mut S,
        _input: &I,
        _exit_kind: &ExitKind
    ) -> Result<(), Error> { ... }
}

Observers observe different information about the target. They can then be used by various sorts of feedback.

Provided Methods

fn flush(&mut self) -> Result<(), Error>

The testcase finished execution, calculate any changes. Reserved for future use.

fn pre_exec(&mut self, _state: &mut S, _input: &I) -> Result<(), Error>

Called right before execution starts.

fn post_exec(
&mut self,
_state: &mut S,
_input: &I,
_exit_kind: &ExitKind
) -> Result<(), Error>

Called right after execution finishes.

fn pre_exec_child(&mut self, _state: &mut S, _input: &I) -> Result<(), Error>

Called right before execution starts in the child process, if any.

fn post_exec_child(
&mut self,
_state: &mut S,
_input: &I,
_exit_kind: &ExitKind
) -> Result<(), Error>

Called right after execution finishes in the child process, if any.