Trait libafl::observers::ObserversTuple

pub trait ObserversTuple<I, S>: MatchName + Debug {
    fn pre_exec_all(&mut self, state: &mut S, input: &I) -> Result<(), Error>;
    fn post_exec_all(
        &mut self,
        state: &mut S,
        input: &I,
        exit_kind: &ExitKind
    ) -> Result<(), Error>;
    fn pre_exec_child_all(
        &mut self,
        state: &mut S,
        input: &I
    ) -> Result<(), Error>;
    fn post_exec_child_all(
        &mut self,
        state: &mut S,
        input: &I,
        exit_kind: &ExitKind
    ) -> Result<(), Error>;
}

A haskell-style tuple of observers

Required Methods

fn pre_exec_all(&mut self, state: &mut S, input: &I) -> Result<(), Error>

This is called right before the next execution.

fn post_exec_all(
&mut self,
state: &mut S,
input: &I,
exit_kind: &ExitKind
) -> Result<(), Error>

This is called right after the last execution

fn pre_exec_child_all(&mut self, state: &mut S, input: &I) -> Result<(), Error>

This is called right before the next execution in the child process, if any.

fn post_exec_child_all(
&mut self,
state: &mut S,
input: &I,
exit_kind: &ExitKind
) -> Result<(), Error>

This is called right after the last execution in the child process, if any.