Trait libafl::schedulers::Scheduler

pub trait Scheduler<I, S> where
    I: Input, {
    fn next(&self, state: &mut S) -> Result<usize, Error>;

    fn on_add(&self, _state: &mut S, _idx: usize) -> Result<(), Error> { ... }
    fn on_replace(
        &self,
        _state: &mut S,
        _idx: usize,
        _prev: &Testcase<I>
    ) -> Result<(), Error> { ... }
    fn on_remove(
        &self,
        _state: &mut S,
        _idx: usize,
        _testcase: &Option<Testcase<I>>
    ) -> Result<(), Error> { ... }
}

The scheduler define how the fuzzer requests a testcase from the corpus. It has hooks to corpus add/replace/remove to allow complex scheduling algorithms to collect data.

Required Methods

fn next(&self, state: &mut S) -> Result<usize, Error>

Gets the next entry

Provided Methods

fn on_add(&self, _state: &mut S, _idx: usize) -> Result<(), Error>

Added an entry to the corpus at the given index

fn on_replace(
&self,
_state: &mut S,
_idx: usize,
_prev: &Testcase<I>
) -> Result<(), Error>

Replaced the given testcase at the given idx

fn on_remove(
&self,
_state: &mut S,
_idx: usize,
_testcase: &Option<Testcase<I>>
) -> Result<(), Error>

Removed the given entry from the corpus at the given index