Trait libafl::stages::mutational::MutationalStage

pub trait MutationalStage<E, EM, I, M, S, Z>: Stage<E, EM, S, Z> where
    M: Mutator<I, S>,
    I: Input,
    S: HasClientPerfMonitor + HasCorpus<I>,
    Z: Evaluator<E, EM, I, S>, {
    fn mutator(&self) -> &M;
    fn mutator_mut(&mut self) -> &mut M;
    fn iterations(&self, state: &mut S, corpus_idx: usize) -> Result<usize, Error>;

    fn perform_mutational(
        &mut self,
        fuzzer: &mut Z,
        executor: &mut E,
        state: &mut S,
        manager: &mut EM,
        corpus_idx: usize
    ) -> Result<(), Error> { ... }
}

A Mutational stage is the stage in a fuzzing run that mutates inputs. Mutational stages will usually have a range of mutations that are being applied to the input one by one, between executions.

Required Methods

fn mutator(&self) -> &M

The mutator registered for this stage

fn mutator_mut(&mut self) -> &mut M

The mutator registered for this stage (mutable)

fn iterations(&self, state: &mut S, corpus_idx: usize) -> Result<usize, Error>

Gets the number of iterations this mutator should run for.

Provided Methods

fn perform_mutational(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
state: &mut S,
manager: &mut EM,
corpus_idx: usize
) -> Result<(), Error>

Runs this (mutational) stage for the given testcase