Module libafl::stages::push

While normal stages call the executor over and over again, push stages turn this concept upside down: A push stage instead returns an iterator that generates a new result for each time it gets called. With the new testcase, you will have to take care about testcase execution, manually. The push stage relies on internal muttability of the supplied Observers.

Re-exports

pub use mutational::StdMutationalPushStage;

Modules

mutational

Mutational stage is the normal fuzzing stage. For the current input, it will perform a range of random mutations, and then run them in the executor.

Structs

PushStageHelper

Helper class for the PushStage trait, taking care of borrowing the shared state

PushStageSharedState

Should be stored inside a [Rc<RefCell<_>>]

Traits

PushStage

A push stage is a generator that returns a single testcase for each call. It’s an iterator so we can chain it. After it has finished once, we will call it agan for the next fuzzer round.