Struct libafl::stages::tmin::MapEqualityFeedback

pub struct MapEqualityFeedback<M> { /* private fields */ }

A feedback which checks if the hash of the currently observed map is equal to the original hash provided

Implementations

impl<M> MapEqualityFeedback<M>

pub fn new(name: &str, obs_name: &str, orig_hash: u64) -> Self

Create a new map equality feedback – can be used with feedback logic