Trait libafl::stages::Stage

pub trait Stage<E, EM, S, Z> {
    fn perform(
        &mut self,
        fuzzer: &mut Z,
        executor: &mut E,
        state: &mut S,
        manager: &mut EM,
        corpus_idx: usize
    ) -> Result<(), Error>;
}

A stage is one step in the fuzzing process. Multiple stages will be scheduled one by one for each input.

Required Methods

fn perform(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
state: &mut S,
manager: &mut EM,
corpus_idx: usize
) -> Result<(), Error>

Run the stage