Module libafl::state

The fuzzer, and state are the core pieces of every good fuzzer

Structs

StdState

The state a fuzz run.

Constants

DEFAULT_MAX_SIZE

The maximum size of a testcase

Traits

HasClientPerfMonitor

Trait for offering a ClientPerfMonitor

HasCorpus

Trait for elements offering a corpus

HasExecutions

Trait for the execution counter

HasMaxSize

Interact with the maximum size

HasMetadata

Trait for elements offering metadata

HasNamedMetadata

Trait for elements offering named metadata

HasRand

Trait for elements offering a rand

HasSolutions

Trait for elements offering a corpus of solutions

HasStartTime

Trait for the starting time

State

The State of the fuzzer. Contains all important information about the current run. Will be used to restart the fuzzing process at any timme.