Struct libafl::state::StdState

pub struct StdState<C, I, R, SC> where
    C: Corpus<I>,
    I: Input,
    R: Rand,
    SC: Corpus<I>, { /* private fields */ }

The state a fuzz run.

Implementations

impl<C, I, R, SC> StdState<C, I, R, SC> where
C: Corpus<I>,
I: Input,
R: Rand,
SC: Corpus<I>,

pub fn load_from_directory<E, EM, Z>(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
manager: &mut EM,
in_dir: &Path,
forced: bool,
loader: &mut dyn FnMut(&mut Z, &mut Self, &Path) -> Result<I, Error>
) -> Result<(), Error> where
Z: Evaluator<E, EM, I, Self>,

Loads inputs from a directory. If forced is true, the value will be loaded, even if it’s not considered to be interesting.

pub fn load_initial_inputs_forced<E, EM, Z>(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
manager: &mut EM,
in_dirs: &[PathBuf]
) -> Result<(), Error> where
Z: Evaluator<E, EM, I, Self>,
EM: EventFirer<I>,

Loads all intial inputs, even if they are not considered interesting. This is rarely the right method, use load_initial_inputs, and potentially fix your Feedback, instead.

pub fn load_initial_inputs<E, EM, Z>(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
manager: &mut EM,
in_dirs: &[PathBuf]
) -> Result<(), Error> where
Z: Evaluator<E, EM, I, Self>,
EM: EventFirer<I>,

Loads initial inputs from the passed-in in_dirs.

impl<C, I, R, SC> StdState<C, I, R, SC> where
C: Corpus<I>,
I: Input,
R: Rand,
SC: Corpus<I>,

pub fn generate_initial_inputs_forced<G, E, EM, Z>(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
generator: &mut G,
manager: &mut EM,
num: usize
) -> Result<(), Error> where
G: Generator<I, Self>,
Z: Evaluator<E, EM, I, Self>,
EM: EventFirer<I>,

Generate num initial inputs, using the passed-in generator and force the addition to corpus.

pub fn generate_initial_inputs<G, E, EM, Z>(
&mut self,
fuzzer: &mut Z,
executor: &mut E,
generator: &mut G,
manager: &mut EM,
num: usize
) -> Result<(), Error> where
G: Generator<I, Self>,
Z: Evaluator<E, EM, I, Self>,
EM: EventFirer<I>,

Generate num initial inputs, using the passed-in generator.

pub fn new<F, O>(
rand: R,
corpus: C,
solutions: SC,
feedback: &mut F,
objective: &mut O
) -> Result<Self, Error> where
F: Feedback<I, Self>,
O: Feedback<I, Self>,

Creates a new State, taking ownership of all of the individual components during fuzzing.